requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
dateparser==1.1.8
decorator==5.1.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
pyarrow==12.0.1
pycryptodome==3.18.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.8.8
six==1.16.0
smmap==5.0.0
streamlit==1.24.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
urllib3==2.0.3
validators==0.20.0
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
